Added Nightmare Governor.
Added the Nightmare Governor. Initial testing is good. Kernel is not set to Nightmare by default, must use Kernel Adiutor or other app to set it. I edited the cpufreq_nightmare.c file (search for "wjh" to see it) due to error. Now resolved. Added some other governors, but they are not functional at this time.